requirements.txt
streamlit>=1.30.0
anthropic>=0.20.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
Pillow>=10.0.0
requests>=2.31.0
